+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Prop `initiallyCollapsed`, to `filter-navigators`, to make filters start out collapsed.
 
 ## [3.34.0] - 2019-10-10
 ### Added